refactor(#1599): split touched over-threshold files (campsite #1116)
- bti.rs (1443→549): extract the point-read decode machinery into bti_point.rs.
- index_reader.rs (1019→803): split into index_reader/{mod,parse}.rs (struct+API vs
  nom parse tree); re-export the parsers by path so callers are unchanged.
